src/core/scale_transition.py
"""Scale Transition System Implementation.

This module implements the scale transition system that connects different scales
in the neural framework, handling:
- Scale transitions between layers
- Pattern scale connections
- Quantum scale bridges
- Scale validation
"""

from dataclasses import dataclass
from typing import List, Tuple, Optional, Dict, Mapping

import torch
import torch.nn as nn
import numpy as np
import torch.nn.functional as F

from .crystal.scale import ScaleSystem
from .quantum.neural_quantum_bridge import NeuralQuantumBridge


@dataclass
class ScaleTransitionConfig:
    """Configuration for scale transitions."""

    min_scale: float = 0.25
    max_scale: float = 4.0
    num_scales: int = 4
    dim: int = 64
    use_quantum_bridge: bool = True
    hidden_dim: int = 64
    dtype: torch.dtype = torch.float32  # Data type for tensors


class ScaleTransitionLayer(nn.Module):
    """Layer that handles transitions between different scales."""
    
    def __init__(self, config: ScaleTransitionConfig):
        super().__init__()
        self.config = config
        
        # Initialize scale system
        self.scale_system = ScaleSystem(
            dim=config.dim,
            num_scales=config.num_scales,
            dtype=config.dtype
        )
        
        # Initialize quantum bridge if enabled
        self.quantum_bridge = (
            NeuralQuantumBridge(
                hidden_dim=config.hidden_dim,
                dtype=config.dtype
            )
            if config.use_quantum_bridge
            else None
        )
        
        # Scale transition networks
        self.scale_up = nn.ModuleList([
            nn.Linear(config.dim, config.dim, dtype=config.dtype)
            for _ in range(config.num_scales - 1)
        ])
        
        self.scale_down = nn.ModuleList([
            nn.Linear(config.dim, config.dim, dtype=config.dtype)
            for _ in range(config.num_scales - 1)
        ])
        
        # Scale normalization
        self.scale_norm = nn.LayerNorm(config.dim, dtype=config.dtype)
    
    def transition_up(
        self, 
        state: torch.Tensor,
        source_scale: float,
        target_scale: float
    ) -> torch.Tensor:
        """Transition to a larger scale."""
        # Validate scales
        if target_scale <= source_scale:
            raise ValueError("Target scale must be larger than source scale")
            
        # Get scale index
        scale_idx = int(np.log2(target_scale / source_scale))
        if scale_idx >= len(self.scale_up):
            raise ValueError("Scale difference too large")
            
        # Apply scale transition
        state = self.scale_up[scale_idx](state)
        state = self.scale_norm(state)
        
        # Apply quantum bridge if enabled
        if self.quantum_bridge is not None:
            state = self.quantum_bridge.bridge_scales(
                state,
                source_scale,
                target_scale
            )
            
        return state
    
    def transition_down(
        self,
        state: torch.Tensor,
        source_scale: float,
        target_scale: float
    ) -> torch.Tensor:
        """Transition to a smaller scale."""
        # Validate scales
        if target_scale >= source_scale:
            raise ValueError("Target scale must be smaller than source scale")
            
        # Get scale index
        scale_idx = int(np.log2(source_scale / target_scale))
        if scale_idx >= len(self.scale_down):
            raise ValueError("Scale difference too large")
            
        # Apply scale transition
        state = self.scale_down[scale_idx](state)
        state = self.scale_norm(state)
        
        # Apply quantum bridge if enabled
        if self.quantum_bridge is not None:
            state = self.quantum_bridge.bridge_scales(
                state,
                source_scale,
                target_scale
            )
            
        return state
    
    def forward(
        self,
        state: torch.Tensor,
        source_scale: float,
        target_scale: float
    ) -> torch.Tensor:
        """Forward pass handling scale transitions in either direction."""
        if target_scale > source_scale:
            return self.transition_up(state, source_scale, target_scale)
        elif target_scale < source_scale:
            return self.transition_down(state, source_scale, target_scale)
        else:
            return state  # No transition needed


class ScaleTransitionSystem:
    """System that manages scale transitions across the neural framework."""
    
    def __init__(self, config: ScaleTransitionConfig):
        self.config = config
        self.transition_layer = ScaleTransitionLayer(config)
        
    def connect_scales(
        self,
        states: List[torch.Tensor],
        scales: List[float]
    ) -> List[torch.Tensor]:
        """Connect states at different scales."""
        if len(states) != len(scales):
            raise ValueError("Number of states must match number of scales")
            
        connected_states = []
        for i, (state, scale) in enumerate(zip(states, scales)):
            # Connect to next scale if not last
            if i < len(states) - 1:
                next_state = self.transition_layer(
                    state,
                    scale,
                    scales[i + 1]
                )
                connected_states.append(next_state)
            else:
                connected_states.append(state)
                
        return connected_states
    
    def connect_pattern_scales(
        self,
        patterns: List[torch.Tensor],
        scales: List[float],
        preserve_symmetry: bool = True
    ) -> List[torch.Tensor]:
        """Connect pattern states across different scales.
        
        Args:
            patterns: List of pattern tensors at different scales
            scales: List of scale factors
            preserve_symmetry: Whether to preserve pattern symmetries
            
        Returns:
            List of connected pattern states
        """
        if len(patterns) != len(scales):
            raise ValueError("Number of patterns must match number of scales")
            
        connected_patterns = []
        for i, (pattern, scale) in enumerate(zip(patterns, scales)):
            if i < len(patterns) - 1:
                # Get scale ratio
                scale_ratio = scales[i + 1] / scale
                
                # Apply scale-aware transition
                next_pattern = self.transition_layer(
                    pattern,
                    scale,
                    scales[i + 1]
                )
                
                if preserve_symmetry:
                    # Preserve pattern amplitudes
                    curr_amp = torch.max(torch.abs(pattern))
                    next_amp = torch.max(torch.abs(next_pattern))
                    amp_ratio = curr_amp / (next_amp + 1e-8)
                    next_pattern = next_pattern * amp_ratio
                    
                    # Preserve wavelengths
                    if scale_ratio > 1:
                        # Upscaling - interpolate to preserve features
                        next_pattern = F.interpolate(
                            next_pattern.unsqueeze(0).unsqueeze(0),
                            scale_factor=scale_ratio,
                            mode='bilinear',
                            align_corners=True
                        ).squeeze(0).squeeze(0)
                    else:
                        # Downscaling - use adaptive pooling to preserve structure
                        target_size = int(pattern.shape[-1] * scale_ratio)
                        next_pattern = F.adaptive_avg_pool2d(
                            next_pattern.unsqueeze(0).unsqueeze(0),
                            (target_size, target_size)
                        ).squeeze(0).squeeze(0)
                
                connected_patterns.append(next_pattern)
            else:
                connected_patterns.append(pattern)
                
        return connected_patterns
    
    def validate_transitions(
        self,
        states: List[torch.Tensor],
        scales: List[float]
    ) -> Mapping[str, torch.Tensor]:
        """Validate scale transitions."""
        if len(states) != len(scales):
            raise ValueError("Number of states must match number of scales")
            
        # Initialize metrics with proper typing
        metrics: Dict[str, torch.Tensor] = {
            "scale_consistency": torch.zeros(len(states) - 1),
            "information_preservation": torch.zeros(len(states) - 1),
        }
        
        if self.config.use_quantum_bridge:
            metrics["quantum_coherence"] = torch.zeros(len(states) - 1)
        
        # Compute metrics between adjacent scales
        for i in range(len(states) - 1):
            # Scale consistency
            metrics["scale_consistency"][i] = torch.norm(
                states[i + 1] - self.transition_layer(
                    states[i],
                    scales[i],
                    scales[i + 1]
                )
            )
            
            # Information preservation
            metrics["information_preservation"][i] = torch.norm(
                torch.svd(states[i])[1] - torch.svd(states[i + 1])[1]
            )
            
            # Quantum coherence if enabled
            if self.config.use_quantum_bridge and self.transition_layer.quantum_bridge is not None:
                metrics["quantum_coherence"][i] = (
                    self.transition_layer.quantum_bridge.compute_coherence(
                        states[i],
                        states[i + 1]
                    )
                )
                
        return metrics
    
    def validate_pattern_transitions(
        self,
        patterns: List[torch.Tensor],
        scales: List[float]
    ) -> Mapping[str, torch.Tensor]:
        """Validate pattern scale transitions.
        
        Args:
            patterns: List of pattern tensors
            scales: List of scale factors
            
        Returns:
            Dictionary of validation metrics
        """
        if len(patterns) != len(scales):
            raise ValueError("Number of patterns must match number of scales")
            
        metrics: Dict[str, torch.Tensor] = {
            "pattern_consistency": torch.zeros(len(patterns) - 1),
            "feature_preservation": torch.zeros(len(patterns) - 1),
            "symmetry_conservation": torch.zeros(len(patterns) - 1)
        }
        
        for i in range(len(patterns) - 1):
            # Pattern consistency across scales
            connected_pattern = self.transition_layer(
                patterns[i],
                scales[i],
                scales[i + 1]
            )
            metrics["pattern_consistency"][i] = torch.norm(
                patterns[i + 1] - connected_pattern
            )
            
            # Feature preservation using spectral analysis
            curr_fft = torch.fft.fft2(patterns[i])
            next_fft = torch.fft.fft2(patterns[i + 1])
            metrics["feature_preservation"][i] = torch.norm(
                torch.abs(curr_fft) - torch.abs(next_fft)
            )
            
            # Symmetry conservation
            curr_symm = torch.max(torch.abs(patterns[i]))
            next_symm = torch.max(torch.abs(patterns[i + 1]))
            metrics["symmetry_conservation"][i] = torch.abs(
                curr_symm - next_symm
            )
            
        return metrics


class ScaleFlowIntegrator(nn.Module):
    """Integrates scale transitions with quantum geometric flow."""
    
    def __init__(self, config: ScaleTransitionConfig):
        super().__init__()
        self.config = config
        
        # Scale-aware quantum components
        self.scale_quantum_proj = nn.Linear(config.dim, config.hidden_dim)
        self.scale_classical_proj = nn.Linear(config.hidden_dim, config.dim)
        
        # Scale-dependent quantum operations
        self.scale_dependent_ops = nn.ModuleList([
            nn.Sequential(
                nn.Linear(config.hidden_dim, config.hidden_dim),
                nn.LayerNorm(config.hidden_dim),
                nn.ReLU()
            )
            for _ in range(config.num_scales)
        ])
        
        # Entanglement tracking
        self.entanglement_tracker = nn.Parameter(
            torch.zeros(config.num_scales, config.hidden_dim)
        )
        
    def compute_scale_quantum_state(
        self,
        state: torch.Tensor,
        scale_idx: int
    ) -> torch.Tensor:
        """Compute scale-dependent quantum state."""
        # Project to quantum space
        quantum_state = self.scale_quantum_proj(state)
        
        # Apply scale-dependent operations
        quantum_state = self.scale_dependent_ops[scale_idx](quantum_state)
        
        # Track entanglement
        self.entanglement_tracker.data[scale_idx] = torch.mean(
            torch.abs(quantum_state), dim=0
        )
        
        return quantum_state
        
    def integrate_flow(
        self,
        state: torch.Tensor,
        source_scale: float,
        target_scale: float,
        quantum_bridge: Optional[NeuralQuantumBridge] = None
    ) -> torch.Tensor:
        """Integrate quantum geometric flow between scales."""
        # Get scale indices
        source_idx = int(np.log2(source_scale / self.config.min_scale))
        target_idx = int(np.log2(target_scale / self.config.min_scale))
        
        # Compute quantum states
        source_quantum = self.compute_scale_quantum_state(state, source_idx)
        
        if quantum_bridge is not None:
            # Use quantum bridge for evolution
            quantum_result = quantum_bridge.neural_to_quantum(source_quantum)
            if isinstance(quantum_result, tuple):
                quantum_state, _ = quantum_result
            else:
                quantum_state = quantum_result
                
            evolved = quantum_bridge.evolve_quantum_state(
                quantum_state,
                time=abs(target_scale - source_scale)
            )
            target_quantum = quantum_bridge.quantum_to_neural(evolved)
        else:
            # Direct evolution in quantum space
            target_quantum = self.compute_scale_quantum_state(state, target_idx)
        
        # Project back to classical space
        evolved_state = self.scale_classical_proj(target_quantum)
        
        return evolved_state
        
    def get_entanglement_metrics(self) -> Dict[str, torch.Tensor]:
        """Get entanglement tracking metrics."""
        return {
            "mean_entanglement": torch.mean(self.entanglement_tracker, dim=1),
            "max_entanglement": torch.max(self.entanglement_tracker, dim=1)[0],
            "entanglement_std": torch.std(self.entanglement_tracker, dim=1)
        }